capture: Don't set linktype explicitly
I'm not completely familiar with libpcap yet, but setting the handle
type results in this error when using `any` as the interface, but not
when the linktype isn't set explicitly.

error setting link type on handle: EN10MB is not one of the DLTs supported by this device

I originally set the value thinking it would ensure consistency with the
assumptions the readers make, but we'll have to revisit that.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>