Introduce capture.Capture interface
This unifies the single and multi capture APIs nicely.

However, the primary motivation is that this allows
creating capture.Handlers after the captures are initialized, but before
the captures are "started", allowing access to configure handlers with
the capture's LinkType before handling packets This removes the
complicated initialization handling for PcapWriterHandler and should
make it easier to implement an PcapNgWriterHandler as well.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>