Do not block in newClientStreamReader
This was causing head of line blocking when a given peer wasn't sending
any packets, preventing the merger from starting.

Instead, create the pcap.Reader on the first call to ReadPacketData,
delaying the read from the underlying io.PipeReader.

However, this means we cannot not know the link type until after we receive
the first packet, so to handle that, send the link type as a grpc header
immediately after the capture is started by the server.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>