Remove capture.MultiCapture
The multi capture design makes it hard to distinguish which packets are
coming from which interface, which is going to be needed to replace the
current pcap reader based approach of handling packets. Instead each
CaptureResponse will send the required information to reconstruct a
gopacket.CaptureInfo directly so we can remove the pcapgo.Reader and
pcapgo.ReaderNg entirely with our own implementation.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>